requirements.txt
tweepy~=3.8
pandas~=0.25
google-cloud-storage~=1.18
python-dateutil~=2.8
pytz==2019.2
requests~=2.22
